. clsids.h, blkalloc.h, dir.cpp, dllexp.cpp, file.cpp, regfs.cpp: moved CFileImage methods from regfs.cpp to file.cpp, and added CDirImage::RemoveFile, and corrected some mistakes.
